payroll process API changes
newly added pay group listing issue fixed, leave request insert issue in me and organization issue fixed, payroll process tab wise API issues fixed, if no data, save and next button insert issue fixed in 3 tabs.

+++ Chef.HRMS.Services/PayrollProcessing/Service/PayrollLeaveDetailsService.cs
@@ -21,36 +21,41 @@ namespace Chef.HRMS.Services
 
         public async Task<int> BulkInsertAsync(List<PayrollLeaveDetails> payrollLeaveDetails)
         {
-            var res = payrollLeaveDetails.Where(x => x.PayrollProcessId > 0).FirstOrDefault();
-            int payrollProcessID = res.PayrollProcessId;
-            int intRet = await DeleteByPayrollProcessID(payrollProcessID);
-            intRet = await payrollComponentDetailsRepository.DeleteByPayrollProcessID(payrollProcessID, 1);
-
-            await payrollLeaveDetailsRepository.BulkInsertAsync(payrollLeaveDetails);
-            List<PayrollComponentDetails> payrollComponent = payrollLeaveDetails.Select(x => new PayrollComponentDetails()
+            int payrollLeaveInsert = 0;
+            if (payrollLeaveDetails.Count !=0)
             {
-                PayrollProcessId = x.PayrollProcessId,
-                PayrollProcessDate = x.PayrollProcessDate,
-                ProcessStatus = x.ProcessStatus,
-                CrAccount = 0,
-                DrAccount = 0,
-                DeductionAmt = 0,
-                DocNum = "",
-                EarningsAmt = 0,
-                EmployeeId = x.EmployeeId,
-                PayrollComponentId = x.LeaveComponentId,
-                CreatedBy = x.CreatedBy,
-                ModifiedBy = x.ModifiedBy,
-                CreatedDate = x.CreatedDate,
-                ModifiedDate = x.ModifiedDate,
-                IsArchived = x.IsArchived,
-                StepNo = 1,
-                EmployeeName = "",
-            }).ToList();
-                await payrollComponentDetailsRepository.BulkInsertAsync(payrollComponent);
+                var res = payrollLeaveDetails.Where(x => x.PayrollProcessId > 0).FirstOrDefault();
+                int payrollProcessID = res.PayrollProcessId;
+                int intRet = await DeleteByPayrollProcessID(payrollProcessID);
+                intRet = await payrollComponentDetailsRepository.DeleteByPayrollProcessID(payrollProcessID, 1);
 
-                return 1;
+                await payrollLeaveDetailsRepository.BulkInsertAsync(payrollLeaveDetails);
+                List<PayrollComponentDetails> payrollComponent = payrollLeaveDetails.Select(x => new PayrollComponentDetails()
+                {
+                    PayrollProcessId = x.PayrollProcessId,
+                    PayrollProcessDate = x.PayrollProcessDate,
+                    ProcessStatus = x.ProcessStatus,
+                    CrAccount = 0,
+                    DrAccount = 0,
+                    DeductionAmt = 0,
+                    DocNum = "",
+                    EarningsAmt = 0,
+                    EmployeeId = x.EmployeeId,
+                    PayrollComponentId = x.LeaveComponentId,
+                    CreatedBy = x.CreatedBy,
+                    ModifiedBy = x.ModifiedBy,
+                    CreatedDate = x.CreatedDate,
+                    ModifiedDate = x.ModifiedDate,
+                    IsArchived = x.IsArchived,
+                    StepNo = 1,
+                    EmployeeName = "",
+                }).ToList();
+                 payrollLeaveInsert = await payrollComponentDetailsRepository.BulkInsertAsync(payrollComponent);
+
+                return payrollLeaveInsert;
             }
+            return payrollLeaveInsert;
+        }
 
         public async Task<int> DeleteByPayrollProcessID(int PayrollProcessID)
         {
